Added min distance when trying to grab
Added min distance when trying to grab,
each time a grab action is tried, we store the distance to the controller for each stick that is falling. If the distance is smaller than the old ones, we keep it. Performance is calculated from the min distance the user was when trying to grab the stick